Move connection manager to separate file
